sparse Hilbert series denominator
we use a map as denominator now and print a sparse version in .out

+++ libnormaliz/simplex.h
@@ -89,7 +89,7 @@ class SimplexEvaluator {
     vector< Integer > TDiag; // diagonal of transpose of generaor matrix after trigonalization
     vector< bool > Excluded;
     vector< Integer > Indicator; 
-    vector<size_t> gen_degrees;
+    vector< long > gen_degrees;
     HilbertSeries Hilbert_Series;
     list< vector<Integer> > Hilbert_Basis;
     list< vector<Integer> > Candidates;
